streamlit/requirements.txt
numpy==1.26.4
Pillow==10.2.0
torch==2.2.1
torchvision==0.17.1
facenet-pytorch==2.5.3
retina-face==0.0.14
opencv-python==4.9.0.80
